ml-server/django/ml_server/requirements.txt
python==3.10.0
pycaret==3.3.2
pandas == 1.5.3
mysqlclient==2.2.4
Django==5.0.6
djangorestframework==3.15.1